Update Code.cpp
The reason for the infinite loop was because if the input was a string instead of an integer the switch would not recognize it. On lines 27 through 32 I put in a loop that checks if the input is an integer and if it’s a string it will set choice to 0 and then go to default.